+++ node-repository/src/main/java/com/yahoo/vespa/hosted/provision/node/NodeAcl.java
@@ -1,7 +1,8 @@
 // Copyright 2016 Yahoo Inc. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
-package com.yahoo.vespa.hosted.provision;
+package com.yahoo.vespa.hosted.provision.node;
 
 import com.google.common.collect.ImmutableList;
+import com.yahoo.vespa.hosted.provision.Node;
 
 import java.util.List;
